requirements.txt
aioredis==1.0.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
cffi==1.13.2
channels==2.1.0
channels-redis==2.1.1
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
Django==2.0.4
django-classy-tags==0.9.0
django-sekizai==1.0.0
djangorestframework==3.8.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.5.6
mysqlclient==1.4.5
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
